pyproject.toml
[tool.poetry]
name = "moneykit"
version = "0.0.0"
authors = ["MoneyKit"]
description = "A client library for accessing MoneyKit API"
license = "MIT"
repository = "https://github.com/moneykit/moneykit-python"


readme = "README.md"
packages = [
    {include = "moneykit"},
]
include = ["CHANGELOG.md", "moneykit/py.typed"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.15.4,<0.25.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"
urllib3 = ">= 1.25.3"

[tool.poetry.group.dev.dependencies]
twine = "^4.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
ruff = ">=0.0.254"
black = ">=23.3.0"
mypy = ">=1.2"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
tag-branch = "main"
strict = true

[tool.poetry-dynamic-versioning.substitution]
files = ["moneykit/__init__.py"]

[tool.ruff]
select = ["I", "E", "W", "F", "C"]
ignore = ["E501", "C901"]
line-length = 120

[tool.black]
line-length = 120
target_version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
